% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_characters}
\alias{fix_characters}
\title{Convert vector of characters to integer or number if needed}
\usage{
fix_characters(item, thetype = "number")
}
\arguments{
\item{item}{string that needs to be converted}

\item{thetype}{string "number" or "integer"}
}
\value{

}
\description{
Convert vector of characters to integer or number if needed
}
