% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_table_by_place.R
\name{generate_html_table_by_place}
\alias{generate_html_table_by_place}
\title{Generate HTML table of data by community}
\usage{
generate_html_table_by_place(the_spreadsheet)
}
\arguments{
\item{the_spreadsheet}{Mass DPH spreadsheet of weekly public health data}
}
\value{

}
\description{
Generate HTML table of data by community
}
\examples{
#'\dontrun{
new_table <- generate_html_table_by_place("Weekly_Dashboard_Data.xlsx")
}
}
