% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cases_category}
\alias{get_cases_category}
\title{Get color for cases per 100K similar to Mass DPH color scheme}
\usage{
get_cases_category(cases)
}
\arguments{
\item{numeric}{cases}
}
\value{
string with case category color
}
\description{
Get color for cases per 100K similar to Mass DPH color scheme
}
