% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_positivity_category}
\alias{get_positivity_category}
\title{Get category color for test positivity rate based on my own scheme}
\usage{
get_positivity_category(positivity)
}
\arguments{
\item{positivity}{numeric}
}
\value{
string with case category color
}
\description{
Get category color for test positivity rate based on my own scheme
}
