% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mwra_sewage_data.R
\name{graph_sewage_data}
\alias{graph_sewage_data}
\title{Generate interactive plotly graph of MWRA sewage testing data}
\usage{
graph_sewage_data(mydf, mytitle = "")
}
\arguments{
\item{mydf}{dataframe of MWRA sewage test data from import_mwra_sewage_data()}

\item{mytitle}{character string of desired graph title}
}
\value{
plotly graph
}
\description{
Data from http://www.mwra.com/biobot/biobotdata.htm generated by import_mwra_sewage_data() function. See mwra-data vignette for details.
}
