% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mwra_sewage_data.R
\name{import_mwra_sewage_data}
\alias{import_mwra_sewage_data}
\title{Import MWRA sewage Covid-19 test data}
\usage{
import_mwra_sewage_data(mwra_sewage_file, api_key = Sys.getenv("pdftable_api"))
}
\arguments{
\item{mwra_sewage_file}{character string name of PDF downloaded from MWRA, or Excel spreadsheet if you already converted the PDF to .xlsx another way.}

\item{api_key}{pdftables API key if using. See https://pdftables.com/pdf-to-excel-api for more details}
}
\value{
data frame with MWRA Covid-19 data samples wrangled for use in ggplot2.
}
\description{
Data from http://www.mwra.com/biobot/biobotdata.htm. See mwra-data vignette for details. Auto conversion of PDF to .xlsx uses the pdftables API (not free). Updated for Nov 2020 format but does not use error bars.
}
