% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macovid_read_weekly_for_map.R
\name{macovid_read_weekly_for_map}
\alias{macovid_read_weekly_for_map}
\title{Read weekly Mass Covid dashboard to join with shapefile for mapping}
\usage{
macovid_read_weekly_for_map(weekly_data_file)
}
\arguments{
\item{weekly_data_file}{Weekly Covid-19 spreadsheet from Mass Dept of Public Health}

\item{gis_data}{sf object available in this package}
}
\value{
sf object for use in making a Leaflet map
}
\description{
See vignette("make-map", package = "WrangleMACovidData") for more info
}
